requirements.txt
python-telegram-bot==20.7
edge-tts==6.1.10
aiohttp==3.9.3
